requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.36
WTForms==3.1.2
email-validator==2.2.0
PyMySQL==1.1.1
python-dotenv==1.0.1
